requirements.txt
google-adk==1.14.0
google-cloud-firestore==2.19.0
google-cloud-pubsub==2.27.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-api-python-client==2.166.0
python-dotenv==1.1.0
uvicorn==0.34.0
google-cloud-bigquery==3.31.0
langchain-community==0.3.27
wikipedia==1.4.0
